clauditor-kut.2: Move SDK seam to _providers/_anthropic.py + ModelResult rename (#144 US-002)
Move every Anthropic-specific SDK helper (transport branches, retry
logic, token extraction, announcement family, constants, exception
classes) from `clauditor._anthropic` into the new canonical home
`clauditor._providers._anthropic`. Rename `AnthropicResult` →
`ModelResult` and add `provider: Literal["anthropic", "openai"] =
"anthropic"` field per DEC-006 of plans/super/144-providers-call-model.md.
Keep `AnthropicResult = ModelResult` as a back-compat alias (same
class object, ``is`` returns True).

`_providers/__init__.py` re-exports the public surface
(`AnthropicHelperError`, `ClaudeCLIError`, `ModelResult`,
`AnthropicResult`, `call_anthropic`, `resolve_transport`) plus updates
`__all__`. The `_auth` re-export must precede `_anthropic` because
the SDK module imports the auth surface at module-load time;
documented inline.

`clauditor._anthropic` reduces to a 110-line back-compat shim that
re-exports every public symbol plus the SDK-seam private surface
(`_sleep`, `_rand_uniform`, `_monotonic`, retry constants, etc.) so
existing call sites (`from clauditor._anthropic import call_anthropic`,
`from clauditor._anthropic import _CLI_AUTO_ANNOUNCEMENT`) keep working
unmodified for one release.

Test patches that target the runtime mutable hooks (`_sleep`,
`_rand_uniform`, `_monotonic`, `_announced_cli_transport`,
`shutil.which`) updated to point at the canonical
`clauditor._providers._anthropic.*` location. Python binding
semantics make these patches' mechanical paths follow the
implementation's location — `from X import Y` creates a fresh
binding that does not propagate `monkeypatch.setattr` to the
canonical module. No test logic, assertions, or class structure
was modified — only the import paths inside `patch(...)` calls.

New `TestModelResult` class adds 5 regression tests:
- `test_model_result_provider_default` — `ModelResult(...)`
  defaults `provider` to `"anthropic"`.
- `test_anthropic_result_alias_identity` —
  `AnthropicResult is ModelResult` (canonical location).
- `test_anthropic_result_alias_identity_via_shim` — the shim
  re-export `clauditor._anthropic.AnthropicResult` is the same
  class object as the canonical `ModelResult`.
- `test_call_anthropic_still_works_via_shim` — shim's
  `call_anthropic` is the same callable object as the canonical
  one.
- `test_provider_field_accepts_openai` — forward-compat check.

NO `call_model` dispatcher in this commit (lands US-003).
NO `.claude/rules/centralized-sdk-call.md` edits (lands US-004).

Quality gates: `uv run ruff check src/ tests/` clean;
`uv run pytest --cov=clauditor` passes 2637 tests at 98.47% coverage
(80% gate enforced).
